feat(typing-effect): add configurable FPS setting and fix delta time … (#7)
* feat(typing-effect): add configurable FPS setting and fix delta time logic

* docs: update README and CHANGELOG for FPS settings

+++ src/typing-effect/ecs/components/lifecycle-component.ts
@@ -1,6 +1,6 @@
 import { MAX_PARTICLES } from "../constants";
 
 export class LifecycleComponent {
-	public life = new Int16Array(MAX_PARTICLES);
-	public maxLife = new Int16Array(MAX_PARTICLES);
+	public life = new Float32Array(MAX_PARTICLES);
+	public maxLife = new Float32Array(MAX_PARTICLES);
 }
